==> scripts/2.2.PCA_sim.Rmd <==
---
title: "PCA simulated data"
author: "Esther Tercero Atencia"
date: "2023-05-17"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
library(tidyverse)
library(ggpubr)
library(plotly)
library(DESeq2)
library(edgeR)
library(here)
```

## In this document we will load the simulated dataset already filtered and plot Principal Component Analysis before and after normalization and correction by possible confounding variables. 
## Some details about the procedure:

> Normalization is performed with vsn approach (code for TMM normalization is also avaiable)
> PCA helps to evaluate any possible effect of the intrinsic properties of the sample (code for 2d and 3d PCA avaiable)
> In simulated dataset we've got only "group" information, but it is performed just to follow the all the steps in the real datasets

```{r}
# filtered counts
filt <- read.csv(here("data/filtered_data/filtered_sim_discovery.csv")) %>% 
  column_to_rownames(var = "GeneID")

# metadaata
metadata <- read.csv(here("data/simulated_data/metadata")) %>% 
  dplyr::select(SampleID, group) %>%
  dplyr::mutate(sample = SampleID,
                group = as.factor(group),
                .keep = "none") %>%
  dplyr::filter(sample %in% colnames(filt))
```


```{r}
### possible confounders
group <- metadata$group
```

```{r}
## pca without normalization and correction
pca <- prcomp(t(filt))
pcamat1 <- as.data.frame(pca$x) %>%
  rownames_to_column(var="sample") %>%
  merge(metadata, by="sample")

var_exp <- round(pca$sdev^2/sum(pca$sdev^2)*100,2)

## split by group
p1 <- ggplot(pcamat1,aes(PC1,PC2,colour=group))+
  geom_point(size=2)+
  theme_bw() +
  labs(
    x = paste("PC1 (", var_exp[1],"%)", sep=""),
    y = paste("PC2 (", var_exp[2],"%)", sep=""),
    title = paste0("PCA splitted by ", "group")
  ) +
    ggsci::scale_color_npg()

pca_2d <- ggarrange(p1)
svg(here("results/PCA/PCA_2d_sim.svg"), width =20, height = 10); pca_2d; dev.off()
```

```{r}
## pca with normalization and correction
##normalization
norm.factors <- DGEList(counts = filt, samples = metadata, group = metadata$group)
norm.factors <- calcNormFactors(norm.factors, method = "TMM")

design = model.matrix(~group) 

norm_counts_vsn <- limma::removeBatchEffect(limma::normalizeVSN(norm.factors), 
                                             design = design) %>% 
   data.frame()
norm_counts_tmm <- limma::removeBatchEffect(cpm(norm.factors, log = TRUE), 
                                             design = design) %>% 
   data.frame()


## change norm WITH THE SPECIFYC NORMALIZATION 
norm <- norm_counts_vsn
#norm <- norm_counts_tmm

## pca normalization and correction
pca <- prcomp(t(norm))
pcamat1 <- as.data.frame(pca$x) %>%
  rownames_to_column(var="sample") %>%
  merge(metadata, by="sample")

var_exp <- round(pca$sdev^2/sum(pca$sdev^2)*100,2)

## split by group
p1 <- ggplot(pcamat1,aes(PC1,PC2,colour=group))+
  geom_point(size=2)+
  theme_bw() +
  labs(
    x = paste("PC1 (", var_exp[1],"%)", sep=""),
    y = paste("PC2 (", var_exp[2],"%)", sep=""),
    title = paste0("PCA splitted by ", "group")
  ) +
    ggsci::scale_color_npg()


#pca_2d <- ggarrange(p1,p2,p3,p4,p5,p6, nrow = 2, ncol =3)
pca_2d <- ggarrange(p1)
pca_3d <- plot_ly(pcamat1, x = ~PC1, y = ~PC2, z = ~PC3, color = ~group, colors = c('#636EFA','#EF553B') ) %>%
  add_markers()

svg(here("results/PCA/PCA_2d_sim_norm_vsn.svg"), width =20, height = 10); pca_2d ; dev.off()

```
