==> scripts/4.2.DE_DESeq2.Rmd <==
---
title: "Differential expression using DESeq2"
author: "Esther Tercero Atencia"
date: "2023-04-25"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(here)
library(tidyverse)
library(ggrepel)
library(DESeq2)
library(edgeR)
library(ggpubr)
library(MASS)
library(dplyr)
source("../utils/functions.R")
```

In this file we include the code necessary to perform DEA using DESeq2 approach. Data loading is also available.

## Data loading

```{r load data}
metadata <- read.csv(here("data/metadata/all_metadata.csv"))

filtered_counts_mm <- read.csv(here("data/filtered_data/filtered_dedupCounts_mm.csv")) %>% 
  column_to_rownames("GeneID")
filtered_counts_mm_discovery <- read.csv(here("data/filtered_data/filtered_mm_discovery.csv")) %>% 
  column_to_rownames("GeneID")

```

## Myeloma multiple DE analysis

```{r}
# metadata
metadata_mm_disco <- metadata %>% 
  filter(sample %in% colnames(filtered_counts_mm_discovery)) %>%
  mutate(sample = sample,
         outcome = as.factor(outcome),
         age = as.numeric(age),
         gender = as.factor(gender),
         RNA_extraction = as.factor(RNA_extraction),
         Library_preparation = as.factor(as.character(Library_preparation)),
         group = as.factor(as.character(group)),
         age_fact = as.factor(cut(age, breaks=4)))
```


```{r}
# correction and normalization using DESeq2 approach
metadata_mm_disco$group <- relevel(metadata_mm_disco$group, ref = "Non-Cancer")
design_all <- model.matrix(~ group + gender + age_fact + Library_preparation, data = metadata_mm_disco) 
colnames(design_all) <- make.names(colnames(design_all))

des_batchremoval <- DESeqDataSetFromMatrix(
  countData = filtered_counts_mm_discovery,  #filtered count matrix
  colData = metadata_mm_disco,  #metadata
  design  = design_all    #model matrix 
)



dataset_batchremoval <- DESeq(des_batchremoval)
results_mm_disco <- results(dataset_batchremoval, cooksCutoff =  F, name = "groupMultiple.Myeloma")

# compute shrunken log2FC
lfc_se_batch <- lfcShrink(
  dataset_batchremoval,
  type = "apeglm",
  res = results_mm_disco,
  coef = "groupMultiple.Myeloma",
  apeAdapt = F)
### resultsNames(dataset_batchremoval)

# arrange results
lfc_se_batch <- dplyr::as_tibble(lfc_se_batch, rownames = "gene") %>%
    dplyr::arrange(pvalue, gene) %>%
    dplyr::transmute(
      gene = gene,
      log2fc = log2FoldChange,
      pvalue = pvalue,
      padj = stats::p.adjust(pvalue, method = "fdr")  # add adjust p-values
      )

# set diffexpressed genes (pvalue < 0.05 and log2fc > 0.5 -in abs value-)
lfc_se_batch$diffexpressed <- "no"
lfc_se_batch$diffexpressed <- ifelse(lfc_se_batch$padj < 0.05 & (lfc_se_batch$log2fc > 0.5 | lfc_se_batch$log2fc < -0.5) , "yes", "no")
lfc_se_batch_mm <- lfc_se_batch

# save results  
write_csv(lfc_se_batch_mm, file = here("results/DESeq2_results_mm_discovery.csv"))

# volcano plot with results
plot_mm <- ggplot(data=lfc_se_batch, aes(x=log2fc, y=-log10(padj), col=diffexpressed)) + 
  geom_point(aes(alpha = diffexpressed)) +
  theme_minimal() +
  scale_color_manual(values=c("gray", "#87CEFA")) +
  scale_alpha_manual(values=c(0.6,0.3)) +
  geom_vline(xintercept=c(-0.5, 0.5), linetype = "dashed") +
  geom_hline(yintercept=-log10(0.05), linetype = "dashed", col="black")+
  ggtitle("DESeq2 in Myeloma Multiple")
ggsave(here("results/DESeq2_plot_mm.png"), plot = plot_mm, device="png", dpi=600, bg = "white")
```


## Simulation analysis

```{r load simulated data}
metadata_simulation <- read.csv(here("data/simulated_data/metadata"))

fcount_simulation_discovery <- read.csv(here("data/filtered_data/filtered_sim_discovery.csv")) %>% 
  column_to_rownames(var="GeneID")

metadata_simulation_discovery <- metadata_simulation %>% 
  filter(SampleID %in% colnames(fcount_simulation_discovery)) %>%
  mutate(group = as.factor(group)) %>%
  mutate(relevel(as.factor(group), ref = "condition"))
```

```{r}
# normalization using DESeq2 approach
design_all <- model.matrix(~ group, data = metadata_simulation_discovery) 
colnames(design_all) <- make.names(colnames(design_all))

des_batchremoval <- DESeqDataSetFromMatrix(
  countData = fcount_simulation_discovery,  #filtered count matrix
  colData = metadata_simulation_discovery,  #metadata
  design  = design_all    #model matrix 
)

dataset_batchremoval <- DESeq(des_batchremoval)
results_simulation_discovery <- results(dataset_batchremoval, name = "groupcontrol")

lfc_se_batch <- lfcShrink(
  dataset_batchremoval,
  type = "apeglm",
  res = results_simulation_discovery,
  coef = "groupcontrol",
  apeAdapt = F)
### resultsNames(dataset_batchremoval)

# arrange results
lfc_se_batch <- dplyr::as_tibble(lfc_se_batch, rownames = "gene") %>%
    dplyr::arrange(pvalue, gene) %>%
    dplyr::transmute(
      gene = gene,
      log2fc = log2FoldChange,
      pvalue = pvalue,
      padj = stats::p.adjust(pvalue, method = "fdr")  # add adjust p-values
      )

# set diffexpressed genes (pvalue < 0.05 and log2fc > 0.5 -in abs value-)
lfc_se_batch$diffexpressed <- "no"
lfc_se_batch$diffexpressed <- ifelse(lfc_se_batch$padj < 0.05 & (lfc_se_batch$log2fc > 0.5 | lfc_se_batch$log2fc < -0.5) , "yes", "no")
lfc_se_batch_sim_discovery <- lfc_se_batch

# save results
write_csv(lfc_se_batch_sim_discovery, file = here("results/DESeq2_results_sim_discovery.csv"))

# volcano plot with results
plot_sim <- ggplot(data=lfc_se_batch, aes(x=log2fc, y=-log10(padj), col=diffexpressed)) + 
  geom_point(aes(alpha = diffexpressed), size = 2) +
  theme_minimal() +
  scale_color_manual(values=c("gray", "#87CEFA")) +
  scale_alpha_manual(values=c(0.6,0.3)) +
  geom_vline(xintercept=c(-0.5, 0.5), linetype = "dashed") +
  geom_hline(yintercept=-log10(0.05), linetype = "dashed", col="black")+
  ggtitle("DESeq2 in simulated dataset")
ggsave(here("results/DESeq2_plot_sim2.png"), plot = plot_sim, device="png", dpi=600, bg = "white", width = 13,
  height = 10, units = "cm")

```
